# Optimize RLE by 40%