--- a/rle/rleplus_reader_test.go
+++ b/rle/rleplus_reader_test.go
@@ -42,3 +42,7 @@ func TestDecodeTable(t *testing.T) {
 		t.Fatalf("not handled")
 	}
 }
+
+func TestAll8bit(t *testing.T) {
+
+}