It is a bit slower for very small RLEs (10 bits or less) but 41% faster for decode+encode
loop with real, big RLE from chain.

Full bench run:
```
Count/0/basic-8              192ns ±10%   218ns ±15%  +13.22%  (p=0.009 n=10+10)
Count/1/basic-8              187ns ±13%   207ns ± 6%  +10.64%  (p=0.001 n=10+9)
Count/1/modified-8           868ns ±11%   962ns ± 6%  +10.77%  (p=0.002 n=10+7)
Count/10/basic-8             797ns ± 6%   823ns ± 6%     ~     (p=0.123 n=10+10)
Count/10/modified-8         1.45µs ± 3%  1.27µs ± 8%  -12.70%  (p=0.000 n=10+10)
Count/1000/basic-8          90.0µs ± 1%  74.5µs ± 7%  -17.29%  (p=0.000 n=9+9)
Count/1000/modified-8       94.5µs ± 0%  82.0µs ± 0%  -13.17%  (p=0.000 n=8+8)
Count/1000000/basic-8        141ms ± 1%   126ms ± 3%  -10.67%  (p=0.000 n=8+9)
Count/1000000/modified-8     146ms ± 1%   128ms ± 3%  -12.16%  (p=0.000 n=9+10)
IsEmpty/0/basic-8            217ns ± 4%   217ns ± 8%     ~     (p=0.591 n=10+10)
IsEmpty/1/basic-8            218ns ± 5%   247ns ±16%  +13.14%  (p=0.000 n=10+10)
IsEmpty/1/modified-8        1.00µs ± 2%  1.04µs ± 6%   +4.03%  (p=0.002 n=7+10)
IsEmpty/10/basic-8           804ns ± 2%   818ns ± 3%     ~     (p=0.161 n=9+9)
IsEmpty/10/modified-8       1.40µs ± 9%  1.45µs ±10%     ~     (p=0.197 n=10+10)
IsEmpty/1000/basic-8        70.9µs ± 1%  70.6µs ± 2%     ~     (p=0.093 n=8+9)
IsEmpty/1000/modified-8     71.7µs ± 1%  61.4µs ± 1%  -14.39%  (p=0.000 n=9+9)
IsEmpty/1000000/basic-8      124ms ± 2%   107ms ± 1%  -13.35%  (p=0.000 n=9+9)
IsEmpty/1000000/modified-8   124ms ± 2%   107ms ± 0%  -13.55%  (p=0.000 n=10+9)
BigDecodeEncode-8           1.37ms ±11%  0.80ms ± 1%  -41.87%  (p=0.000 n=10+9)
```