@Kubuxu 2020-11-03
@Stebalien is the test testing all values in the decode table enough? or do you want the reader to actually decode these runs?