@Stebalien 2020-11-03
I'm more interested in actually decoding them. I trust we'll generate the table correctly, I just want to make sure correctly is the correct correctly.